Move show referral integration tests
- Add `sharedTests` to provide the body of tests that are repeated in
  different locations. The immediate use for this is the show referral
  test bodies, which will be used in Assess and Refer contexts
- Use `sharedTests` in Refer
